move test into test_subscription
